LPS-84119 Move logic to enforce correct version to new class GradleDependencyArtifactsCheck
