+++ modules/source-formatter.properties
@@ -5,7 +5,7 @@
 ## for available properties.
 ##
 
-    source.check.gradle.dependency.version.check.enforceVersionArtifacts=\
+    source.check.gradle.dependency.artifacts.check.enforceVersionArtifacts=\
         biz.aQute.bnd:biz.aQute.bnd.annotation:4.2.0,\
         com.fasterxml.jackson.core:jackson-annotations:2.9.9,\
         com.fasterxml.jackson.core:jackson-core:2.9.9,\
