+++ source-formatter.properties
@@ -252,7 +252,7 @@
 
     source.check.gradle.dependencies.check.checkPetraDependencies=true
 
-    source.check.gradle.dependency.version.check.enforceVersionArtifacts=
+    source.check.gradle.dependency.artifacts.check.enforceVersionArtifacts=
 
     source.check.html.empty.lines.check.enabled=true
